// server/prisma/__generated__/graphql/resolvers/inputs/StringNullableWithAggregatesFilter.ts
import * as TypeGraphQL from "type-graphql";
import * as GraphQLScalars from "graphql-scalars";
import { Prisma } from "@prisma/client";
import { DecimalJSScalar } from "../../scalars";
import { NestedIntNullableFilter } from "../inputs/NestedIntNullableFilter";
import { NestedStringNullableFilter } from "../inputs/NestedStringNullableFilter";
import { NestedStringNullableWithAggregatesFilter } from "../inputs/NestedStringNullableWithAggregatesFilter";

@TypeGraphQL.InputType("StringNullableWithAggregatesFilter", {})
export class StringNullableWithAggregatesFilter {
    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    equals?: string | undefined;

    @TypeGraphQL.Field(_type => [String], {
            nullable: true
        })
    in?: string[] | undefined;

    @TypeGraphQL.Field(_type => [String], {
            nullable: true
        })
    notIn?: string[] | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    lt?: string | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    lte?: string | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    gt?: string | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    gte?: string | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    contains?: string | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    startsWith?: string | undefined;

    @TypeGraphQL.Field(_type => String, {
            nullable: true
        })
    endsWith?: string | undefined;

    @TypeGraphQL.Field(_type => NestedStringNullableWithAggregatesFilter, {
            nullable: true
        })
    not?: NestedStringNullableWithAggregatesFilter | undefined;

    @TypeGraphQL.Field(_type => NestedIntNullableFilter, {
            nullable: true
        })
    _count?: NestedIntNullableFilter | undefined;

    @TypeGraphQL.Field(_type => NestedStringNullableFilter, {
            nullable: true
        })
    _min?: NestedStringNullableFilter | undefined;

    @TypeGraphQL.Field(_type => NestedStringNullableFilter, {
            nullable: true
        })
    _max?: NestedStringNullableFilter | undefined;
}